tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-radius-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing border-radius with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-radius">
<meta name="assert" content="border-radius supports only the grammar '<length-percentage>{1,4} [ / <length-percentage>{1,4} ]?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("border-radius", "none");
test_invalid_value("border-radius", "1px 2px 3px 4px 5px");
test_invalid_value("border-radius", "-1px");
test_invalid_value("border-radius", "1px -2px");
test_invalid_value("border-radius", "1em /");
test_invalid_value("border-radius", "1px / 2px / 3px");

test_invalid_value("border-radius", "5em 1px 5em 2% 5em 3px 5em 4%");
test_invalid_value("border-radius", "1px 5em 2% 5em 3px 5em 4% 5em");

test_invalid_value("border-top-left-radius", "10px 20px 30px");
</script>
</body>
</html>
